pyproject.toml
[project]
name = "chatbot-wshap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "openai>=2.36.0",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]
